feat(Layout): Implement Layout component to wrap Header and Speakers components
